LayoutTests/imported/w3c/web-platform-tests/css/css-scrollbars/scrollbar-width-001-expected.txt

Test scrollbar-width: vertical scrollbar, ltr direction
auto
thin
none
Test scrollbar-width: vertical scrollbar, rtl direction
auto
thin
none

PASS horizontal-tb, ltr, scrollbar-width auto
PASS horizontal-tb, ltr, scrollbar-width thin
PASS horizontal-tb, ltr, scrollbar-width "thin" is same or thinner than "auto"
FAIL horizontal-tb, ltr, scrollbar-width none assert_equals: ltr none scrollWidth expected 200 but got 185
PASS horizontal-tb, rtl, scrollbar-width auto
PASS horizontal-tb, rtl, scrollbar-width thin
PASS horizontal-tb, rtl, scrollbar-width "thin" is same or thinner than "auto"
FAIL horizontal-tb, rtl, scrollbar-width none assert_equals: rtl none scrollWidth expected 200 but got 185
